feat: configure application properties and add PostgreSQL support with Docker

+++ src/main/java/com/readtrack/userservice/infrastructure/repositories/UserRepository.java
@@ -3,7 +3,7 @@ package com.readtrack.userservice.infrastructure.repositories;
 import com.readtrack.userservice.infrastructure.models.UserMO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserMO, Integer> {
+public interface UserRepository extends JpaRepository<UserMO, Long> {
     boolean existsByUsername(String username);
 
     boolean existsByEmail(String email);
